Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "dotenv>=0.9.9",
    "email-validator>=2.3.0",
    "fastapi>=0.121.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
